requirements.txt
absl-py==2.1.0
casadi==3.6.4
clarabel==0.9.0
cloudpickle==3.1.0
contourpy==1.3.1
cvxpy==1.4.1
cycler==0.12.1
dill==0.3.7
ecos==2.0.14
Farama-Notifications==0.0.4
ffmpeg-python==0.2.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
grpcio==1.67.1
gymnasium==0.29.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpi4py==4.0.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.5
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
psutil==6.1.0
pybind11==2.13.6
pybullet==3.0.0
pygame==2.5.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
qdldl==0.1.7.post4
scipy==1.14.1
scs==3.2.7
six==1.16.0
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.2.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3